+++ packages/constructor/src/constructors/FormattedNumber/index.ts
@@ -1,0 +1,5 @@
+export * from './FormattedNumber'
+export * from './FormattedNumberDesign'
+export * from './props'
+export * from './types'
+export * from './basicTypes'