+++ packages/d1/src/components/Ui/FormattedNumber/wikiData.ts
@@ -1,0 +1,33 @@
+import { type StorybookProps, type StorybookSlots, type WikiDataItem } from '@dxtmisha/wiki'
+
+import { wikiD1 } from '../../../wiki/wiki'
+import { defaults } from './props'
+
+const propsNames: StorybookProps = [
+  // :propsList [!] System label / Системная метка
+  { name: 'currency', type: 'string' },
+  { name: 'format', type: 'boolean' },
+  { name: 'language', type: 'string' },
+  { name: 'unit', type: 'string' },
+  { name: 'value', type: 'NumberOrString' }
+  // :propsList [!] System label / Системная метка
+]
+
+const slotsNames: StorybookSlots = [
+  // :slotsList [!] System label / Системная метка
+  // :slotsList [!] System label / Системная метка
+]
+
+const eventsNames: StorybookSlots = [
+  // :eventsList [!] System label / Системная метка
+  // :eventsList [!] System label / Системная метка
+]
+
+export const FormattedNumberWikiData: WikiDataItem = {
+  component: 'FormattedNumber',
+  props: propsNames,
+  slots: slotsNames,
+  events: eventsNames,
+  defaults,
+  wikiDesign: wikiD1
+}